requirements.txt
python-telegram-bot==20.7
Flask==3.0.0
APScheduler==3.10.4
requests
